implements task 2273 "Move all custom msft visuals to Github" creates a new project "CustomVisuals"
