+++ src/Clients/CustomVisuals/_references.ts
@@ -1,4 +1,4 @@
-﻿/*
+/*
  *  Power BI Visualizations
  *
  *  Copyright (c) Microsoft Corporation
@@ -11,10 +11,10 @@
  *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  *  copies of the Software, and to permit persons to whom the Software is
  *  furnished to do so, subject to the following conditions:
- *   
+ *
  *  The above copyright notice and this permission notice shall be included in 
  *  all copies or substantial portions of the Software.
- *   
+ *
  *  THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR 
  *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, 
  *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE 
@@ -24,7 +24,25 @@
  *  THE SOFTWARE.
  */
 
-/// <reference path="../../VisualsCommon/obj/VisualsCommon.d.ts" />
-/// <reference path="../../VisualsData/obj/VisualsData.d.ts" />
-/// <reference path="../../Visuals/obj/Visuals.d.ts" />
-/// <reference path="../../CustomVisuals/obj/CustomVisuals.d.ts" />
+/// <reference path="../Visuals/obj/visuals.d.ts" />
+
+/// <reference path="visuals/consoleWriter/consoleWriter.ts"/>
+/// <reference path="visuals/helloIVisual/helloIVisual.ts"/>
+/// <reference path="visuals/asterPlot/asterPlot.ts"/>
+/// <reference path="visuals/streamGraph/streamGraph.ts"/>
+/// <reference path="visuals/sunburst/sunburst.ts"/>
+/// <reference path="visuals/radarChart/radarChart.ts"/>
+/// <reference path="visuals/dotPlot/dotPlot.ts"/>
+/// <reference path="visuals/timeline/timeline.ts"/>
+/// <reference path="visuals/areaRangeChart/areaRangeChart.ts"/>
+/// <reference path="visuals/chicletSlicer/chicletSlicer.ts"/>
+/// <reference path="visuals/tornadoChart/tornadoChart.ts" />
+/// <reference path="visuals/histogram/histogram.ts"/>
+/// <reference path="visuals/sankeyDiagram/sankeyDiagram.ts"/>
+/// <reference path="visuals/wordCloud/wordCloud.ts"/>
+/// <reference path="visuals/enhancedScatterChart/enhancedScatterChart.ts" />
+/// <reference path="visuals/gantt/gantt.ts" />
+/// <reference path="visuals/globeMap/globeMap.ts" />
+/// <reference path="visuals/bulletChart/bulletChart.ts" />
+
+/// <reference path="plugins.ts" />
