+++ src/Clients/CustomVisuals/visuals/areaRangeChart/areaRangeChart.ts
