+++ src/Clients/CustomVisuals/visuals/chicletSlicer/styles/chicletSlicer.less
@@ -1,0 +1,126 @@
+/*
+ *  Power BI Visualizations
+ *
+ *  Copyright (c) Microsoft Corporation
+ *  All rights reserved. 
+ *  MIT License
+ *
+ *  Permission is hereby granted, free of charge, to any person obtaining a copy
+ *  of this software and associated documentation files (the ""Software""), to deal
+ *  in the Software without restriction, including without limitation the rights
+ *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ *  copies of the Software, and to permit persons to whom the Software is
+ *  furnished to do so, subject to the following conditions:
+ *
+ *  The above copyright notice and this permission notice shall be included in 
+ *  all copies or substantial portions of the Software.
+ *
+ *  THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR 
+ *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, 
+ *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE 
+ *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER 
+ *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ *  THE SOFTWARE.
+ */
+
+@import (reference) "../../../styles/styles.less";
+
+.chicletSlicer {
+    font-family: @regularFontFamily;
+
+    .clear {
+        background-image: url('data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAA8AAAAOCAIAAAB/6NG4AAAAAXNSR0IArs4c6QAAAARnQU1BAACxjwv8YQUAAAAJcEhZcwAAHYcAAB2HAY/l8WUAAADtSURBVChTlZBLioRAEETn/jdwJSjqThFBRRARQRD8LjyCCwUPoKJOVEf1tC09MPNWFVkvsz5f53/41c6yrKoqGZ58tqFGUeR5XtM0svTggw01DMN937dtcxzn2nC3i6KgyrgsCxq6rmN8s28qYUPf91i/bKi2beN0mS/keY4pWEi7rmvXdQ3DGIaBlR8wJQgCHijstm3xfEyd59k0zXEcH5oAKrcYha2qKi7HjAbLsthQluVVBcJO0zSO4+M4WJqmCVfCP95UIGx4sJMkYQlAxT+s6yrzE/lKNui6jqlAURRN07DwfZ8CkfafOM9vHlBGJTazbYkAAAAASUVORK5CYII=');
+
+        background-size: contain;
+        background-repeat: no-repeat;
+        background-position: center;
+        width: 13px;
+        height: 12px;
+        margin: 4px 2px;
+        float: right;
+        cursor: pointer;
+    }
+
+    .row {
+      white-space: nowrap;
+    }
+
+    .hidden {
+      display: none !important;
+    }
+
+    .slicerBody {
+        .cell {
+            box-sizing: border-box;
+            padding: 2px 5px;
+
+            display: inline-block;
+
+            position: relative;
+            cursor: pointer;
+            width: 100%;
+        }
+
+        li.slicerItemContainer {
+            height: 100%;
+            width: 100%;
+            padding: 2px;
+            display: flex;
+            flex-direction: column;
+            flex-wrap: nowrap;
+            justify-content: center;
+            align-content: center;
+            align-items: stretch;
+            border-radius: 5px;
+            box-sizing: border-box;
+
+            .slicer-text-wrapper {
+                box-sizing: border-box;
+                text-align: center;
+                order: 0;
+                align-self: auto;
+                flex-grow: 1;
+                flex-shrink: 1;
+                flex-basis: auto;
+                display: flex;
+                flex-direction: column;
+                flex-wrap: nowrap;
+                justify-content: center;
+                align-content: center;
+                align-items: stretch;
+                overflow: hidden;
+                text-overflow: ellipsis;
+            }
+            .slicer-img-wrapper {
+                background-repeat: no-repeat;
+                background-position: center;
+                background-size: contain;
+            }
+            .stretchImage {
+                background-size: 100% 100%;
+            }
+            .bottomImage {
+                order: 1;
+            }
+            .slicerItem-disabled {
+                opacity: 0.4;
+                color: #fff !important;
+            }
+        }
+    }
+
+    .slicerItemContainer.selected {
+        background: #BDD7EE;
+    }
+
+    .row-horizontal {
+        display: inline-block;
+    }
+    .slicerContainer .slicerText {
+        padding: 0px 5px;
+    }
+    .slicerHeader {
+        margin-bottom: 5px;
+    }
+}
