+++ src/Clients/CustomVisuals/visuals/consoleWriter/consoleWriter.ts
