+++ src/Clients/CustomVisuals/visuals/globeMap/globeMap.ts
