+++ src/Clients/CustomVisuals/visuals/histogram/histogram.ts
