+++ src/Clients/CustomVisuals/visuals/streamGraph/styles/streamGraph.less
@@ -1,4 +1,4 @@
-﻿/*
+/*
  *  Power BI Visualizations
  *
  *  Copyright (c) Microsoft Corporation
@@ -11,10 +11,10 @@
  *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  *  copies of the Software, and to permit persons to whom the Software is
  *  furnished to do so, subject to the following conditions:
- *   
+ *
  *  The above copyright notice and this permission notice shall be included in 
  *  all copies or substantial portions of the Software.
- *   
+ *
  *  THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR 
  *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, 
  *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE 
@@ -24,7 +24,23 @@
  *  THE SOFTWARE.
  */
 
-/// <reference path="../../VisualsCommon/obj/VisualsCommon.d.ts" />
-/// <reference path="../../VisualsData/obj/VisualsData.d.ts" />
-/// <reference path="../../Visuals/obj/Visuals.d.ts" />
-/// <reference path="../../CustomVisuals/obj/CustomVisuals.d.ts" />
+@import (reference) "../../../styles/styles.less";
+
+.streamGraph {
+    @streamGraphFontSize: 11px;
+
+    font-family: @regularFontFamily;
+    font-size: @streamGraphFontSize;
+
+    .axis {
+        path, line {
+            fill: none;
+            stroke: #666;
+            shape-rendering: @defaultShapeRendering;
+        }
+
+        text {
+            font-size: @streamGraphFontSize;
+        }
+    }
+}
