+++ src/Clients/CustomVisuals/visuals/tornadoChart/styles/tornadoChart.less
@@ -1,4 +1,4 @@
-﻿/*
+/*
  *  Power BI Visualizations
  *
  *  Copyright (c) Microsoft Corporation
@@ -11,10 +11,10 @@
  *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  *  copies of the Software, and to permit persons to whom the Software is
  *  furnished to do so, subject to the following conditions:
- *   
+ *
  *  The above copyright notice and this permission notice shall be included in 
  *  all copies or substantial portions of the Software.
- *   
+ *
  *  THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR 
  *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, 
  *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE 
@@ -24,7 +24,19 @@
  *  THE SOFTWARE.
  */
 
-/// <reference path="../../VisualsCommon/obj/VisualsCommon.d.ts" />
-/// <reference path="../../VisualsData/obj/VisualsData.d.ts" />
-/// <reference path="../../Visuals/obj/Visuals.d.ts" />
-/// <reference path="../../CustomVisuals/obj/CustomVisuals.d.ts" />
+@import (reference) "../../../styles/styles.less";
+
+.tornado-chart {
+    @tornadoChartAxisColor: @neutralSecondaryAltColor;
+
+    font-family: @regularFontFamily;
+    font-size: 11px;
+
+    .axes {
+        .axis {
+            fill: none;
+            stroke: @tornadoChartAxisColor;
+            stroke-width: 1px;
+        }
+    }
+}
