+++ src/Clients/PowerBIVisualsPlayground/.gitignore
@@ -1,1 +1,2 @@
 images/visuals.sprites.png
+CustomVisuals.css
