+++ src/Clients/PowerBIVisualsPlayground/index.html
@@ -6,9 +6,11 @@
     <title>Power BI visualization playground</title>
     <link rel="stylesheet" href="app.css" />
     <link rel="stylesheet" href="visuals.css" />
+    <link rel="stylesheet" href="CustomVisuals.css" />
 
-    <script type="text/javascript" src="externals.min.js"></script>  
-    <script type="text/javascript" src="powerbi-visuals.js"></script>    
+    <script type="text/javascript" src="externals.min.js"></script>
+    <script type="text/javascript" src="powerbi-visuals.js"></script>
+    <script type="text/javascript" src="CustomVisuals.js"></script>
     <script type="text/javascript" src ="PowerBIVisualsPlayground.js"></script>
     
 </head>
